requirements.txt
-i https://pypi.org/simple
alembic==1.6.5
asgiref==3.4.1; python_version >= '3.6'
bcrypt==3.2.0
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.1; python_version >= '3.6'
cryptography==3.4.7
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
cssutils==2.3.0; python_version >= '3.6'
dnspython==2.1.0; python_version >= '3.6'
ecdsa==0.17.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.1.3
emails==0.6
fastapi-sqlalchemy==0.2.1
fastapi==0.66.0
google-auth==1.32.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
greenlet==1.1.0; python_version >= '3'
h11==0.12.0; python_version >= '3.6'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.1; python_version >= '3.6'
oauthlib==3.1.1; python_version >= '3.6'
passlib[bcrypt]==1.7.4
premailer==3.9.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic[email]==1.8.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7.2; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
starlette==0.14.2; python_version >= '3.6'
typing-extensions==3.10.0.0
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.14.0
websocket-client==1.1.0; python_version >= '3.6'
